Break out Gradio controllers into utils package
