+++ server/utils/setup.py
@@ -2,6 +2,6 @@ from setuptools import setup, find_packages
 
 setup(
     name="utils",  # Name of your module
-    version="0.0.1",
+    version="0.0.2",
     packages=find_packages(),  # Automatically find all packages
 )
